# backend/docs/swagger.yaml
definitions:
...
  model.ChatMessage:
    properties:
      circle_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      from_user:
        $ref: '#/definitions/model.ChatUserInfo'
      id:
        type: string
      to_user_id:
        type: string
      type:
        type: string
    type: object
  model.ChatMessagesResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/model.ChatMessage'
        type: array
      pagination:
        $ref: '#/definitions/model.PaginationResponse'
    type: object
  model.ChatUserInfo:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      code:
        type: integer
      detail:
        type: string
      message:
        type: string
    type: object
...
  model.OnlineUser:
    properties:
      avatar:
        type: string
      is_online:
        type: boolean
      last_seen:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
  model.OnlineUsersResponse:
    properties:
      online_users:
        items:
          $ref: '#/definitions/model.OnlineUser'
        type: array
    type: object
  model.PaginationResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  model.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      timestamp:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /chat/messages:
    get:
      consumes:
      - application/json
      description: 获取私聊或圈子聊天记录
      parameters:
      - description: 目标用户ID（私聊时使用）
        in: query
        name: target_id
        type: string
      - description: 圈子ID（圈子聊天时使用）
        in: query
        name: circle_id
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 50
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.ChatMessagesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: 获取聊天记录
      tags:
      - 聊天
  /chat/online-users:
    get:
      consumes:
      - application/json
      description: 获取当前在线用户列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.OnlineUsersResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: 获取在线用户
      tags:
      - 聊天
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"